+++ src/display_impl.rs
@@ -41,6 +41,7 @@ impl Display for InputKind {
             InputKind::DualAxis(axis) => write!(f, "{axis:?}"),
             InputKind::GamepadButton(button) => write!(f, "{button:?}"),
             InputKind::Mouse(button) => write!(f, "{button:?}"),
+            InputKind::MouseWheel(button) => write!(f, "{button:?}"),
             InputKind::Keyboard(button) => write!(f, "{button:?}"),
         }
     }